+++ src/main/kotlin/io/openfuture/chain/core/service/block/DefaultBlockService.kt
@@ -31,6 +31,11 @@ class DefaultBlockService(
     override fun getLast(): Block =
         repository.findFirstByOrderByHeightDesc() ?: throw NotFoundException("Last block not found!")
 
+    @Transactional
+    override fun save(block: Block) {
+        repository.save(block)
+    }
+
     @Transactional(readOnly = true)
     override fun isExists(hash: String): Boolean = repository.findOneByHash(hash)?.let { true } ?: false
 
@@ -40,11 +45,5 @@ class DefaultBlockService(
 
     @Transactional(readOnly = true)
     override fun getCurrentHeight(): Long = repository.getCurrentHeight()
-
-    @Transactional(readOnly = true)
-    override fun getCarcassForBlockSync(hash: String): List<Block> {
-        val startBlock = repository.findOneByHash(hash) ?: return emptyList()
-        return repository.findTop30ByHeightGreaterThanOrderByHeightDesc(startBlock.height)
-    }
 
 }