+++ src/main/kotlin/io/openfuture/chain/core/service/block/DefaultGenesisBlockService.kt
@@ -12,7 +12,6 @@ import io.openfuture.chain.core.service.DefaultDelegateService
 import io.openfuture.chain.core.service.GenesisBlockService
 import io.openfuture.chain.core.service.WalletService
 import io.openfuture.chain.core.sync.BlockchainLock
-import io.openfuture.chain.core.sync.SyncManager
 import io.openfuture.chain.crypto.util.SignatureUtils
 import io.openfuture.chain.network.message.sync.GenesisBlockMessage
 import io.openfuture.chain.rpc.domain.base.PageRequest
@@ -26,14 +25,10 @@ class DefaultGenesisBlockService(
     blockService: BlockService,
     walletService: WalletService,
     delegateService: DefaultDelegateService,
-    repository: GenesisBlockRepository,
     private val keyHolder: NodeKeyHolder,
-    private val syncManager: SyncManager,
-    private val consensusProperties: ConsensusProperties
-) : BaseBlockService<GenesisBlock>(repository, blockService, walletService, delegateService), GenesisBlockService {
-
-    @Volatile private var last: GenesisBlock = repository.findFirstByOrderByHeightDesc()!!
-
+    private val consensusProperties: ConsensusProperties,
+    private val genesisBlockRepository: GenesisBlockRepository
+) : BaseBlockService<GenesisBlock>(genesisBlockRepository, blockService, walletService, delegateService), GenesisBlockService {
 
     @Transactional(readOnly = true)
     override fun getPreviousByHeight(height: Long): GenesisBlock = repository.findFirstByHeightLessThanOrderByHeightDesc(height)
@@ -61,12 +56,15 @@ class DefaultGenesisBlockService(
     @Transactional(readOnly = true)
     override fun getAll(request: PageRequest): Page<GenesisBlock> = repository.findAll(request)
 
+    override fun getLast(): GenesisBlock = genesisBlockRepository.findFirstByOrderByHeightDesc()!!
+
     @Transactional(readOnly = true)
-    override fun getLast(): GenesisBlock = last
+    override fun getByEpochIndex(epochIndex: Long): GenesisBlock? =
+        genesisBlockRepository.findOneByPayloadEpochIndex(epochIndex)
 
     @BlockchainSynchronized
     @Transactional
-    override fun create(): GenesisBlockMessage {
+    override fun create(): GenesisBlock {
         val lastBlock = blockService.getLast()
         val timestamp = getTimestamp()
         val height = lastBlock.height + 1
@@ -76,8 +74,13 @@ class DefaultGenesisBlockService(
         val signature = SignatureUtils.sign(hash, keyHolder.getPrivateKey())
         val publicKey = keyHolder.getPublicKeyAsHexString()
 
-        return GenesisBlockMessage(height, previousHash, timestamp, ByteUtils.toHexString(hash), signature,
-            publicKey, payload.epochIndex, payload.activeDelegates.map { it.publicKey })
+        return GenesisBlock(timestamp, height, previousHash, ByteUtils.toHexString(hash), signature, publicKey, payload)
+    }
+
+    @Transactional
+    @Synchronized
+    override fun add(block: GenesisBlock) {
+        super.save(block)
     }
 
     @Transactional
@@ -90,20 +93,14 @@ class DefaultGenesisBlockService(
         val delegates = message.delegates.asSequence().map { delegateService.getByPublicKey(it) }.toMutableList()
         val block = GenesisBlock.of(message, delegates)
 
-        if (!isSync(block)) {
-            syncManager.outOfSync()
-            return
-        }
-
-        super.save(block)
-        last = block
+        add(block)
     }
 
     override fun isGenesisBlockRequired(): Boolean {
         BlockchainLock.readLock.lock()
         try {
             val blocksProduced = blockService.getCurrentHeight() - getLast().height
-            return (consensusProperties.epochHeight!! - 1) <= blocksProduced
+            return (consensusProperties.epochHeight!!) <= blocksProduced
         } finally {
             BlockchainLock.readLock.unlock()
         }