+++ src/main/kotlin/io/openfuture/chain/network/handler/sync/GenesisBlockHandler.kt
@@ -4,7 +4,6 @@ import io.netty.channel.ChannelHandler.Sharable
 import io.netty.channel.ChannelHandlerContext
 import io.netty.channel.SimpleChannelInboundHandler
 import io.openfuture.chain.core.service.GenesisBlockService
-import io.openfuture.chain.core.sync.SyncManager
 import io.openfuture.chain.network.message.sync.GenesisBlockMessage
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
@@ -13,7 +12,6 @@ import org.springframework.stereotype.Component
 @Component
 @Sharable
 class GenesisBlockHandler(
-    private val syncManager: SyncManager,
     private val genesisBlockService: GenesisBlockService
 ) : SimpleChannelInboundHandler<GenesisBlockMessage>() {
 
@@ -24,9 +22,7 @@ class GenesisBlockHandler(
 
     override fun channelRead0(ctx: ChannelHandlerContext, msg: GenesisBlockMessage) {
         log.debug("GenesisBlockHandler: ${msg.height}-${msg.hash} from ${ctx.channel().remoteAddress()}")
-        syncManager.onBlockMessage(msg) {
-            genesisBlockService.add(msg)
-        }
+        genesisBlockService.add(msg)
     }
 
 }