+++ src/main/kotlin/io/openfuture/chain/network/handler/sync/MainBlockHandler.kt
@@ -4,7 +4,6 @@ import io.netty.channel.ChannelHandler.Sharable
 import io.netty.channel.ChannelHandlerContext
 import io.netty.channel.SimpleChannelInboundHandler
 import io.openfuture.chain.core.service.MainBlockService
-import io.openfuture.chain.core.sync.SyncManager
 import io.openfuture.chain.network.message.sync.MainBlockMessage
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
@@ -13,7 +12,6 @@ import org.springframework.stereotype.Component
 @Component
 @Sharable
 class MainBlockHandler(
-    private val syncManager: SyncManager,
     private val mainBlockService: MainBlockService
 ) : SimpleChannelInboundHandler<MainBlockMessage>() {
 
@@ -24,9 +22,7 @@ class MainBlockHandler(
 
     override fun channelRead0(ctx: ChannelHandlerContext, msg: MainBlockMessage) {
         log.debug("MainBlockHandler: ${msg.height}-${msg.hash} from ${ctx.channel().remoteAddress()}")
-        syncManager.onBlockMessage(msg) {
-            mainBlockService.add(msg)
-        }
+        mainBlockService.add(msg)
     }
 
 }