+++ src/main/kotlin/io/openfuture/chain/network/handler/sync/SyncResponseHandler.kt
@@ -3,11 +3,8 @@ package io.openfuture.chain.network.handler.sync
 import io.netty.channel.ChannelHandler.Sharable
 import io.netty.channel.ChannelHandlerContext
 import io.netty.channel.SimpleChannelInboundHandler
-import io.openfuture.chain.core.sync.SyncManager
-import io.openfuture.chain.network.component.ChannelsHolder
-import io.openfuture.chain.network.component.time.Clock
+import io.openfuture.chain.core.sync.ChainSynchronizer
 import io.openfuture.chain.network.message.sync.SyncResponseMessage
-import io.openfuture.chain.network.property.NodeProperties
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Component
@@ -15,26 +12,15 @@ import org.springframework.stereotype.Component
 @Component
 @Sharable
 class SyncResponseHandler(
-    private val clock: Clock,
-    private val syncManager: SyncManager,
-    private val properties: NodeProperties
+    private val chainSynchronizer: ChainSynchronizer
 ) : SimpleChannelInboundHandler<SyncResponseMessage>() {
 
     companion object {
         private val log: Logger = LoggerFactory.getLogger(SyncResponseHandler::class.java)
     }
-
-
     override fun channelRead0(ctx: ChannelHandlerContext, msg: SyncResponseMessage) {
-        if (properties.expiry!! < clock.currentTimeMillis() - msg.timestamp) {
-            log.debug("Expired sync response")
-            return
-        }
-
-        log.debug("RESPONSE from ${ctx.channel().remoteAddress()}")
-        ctx.channel().attr(ChannelsHolder.NODE_INFO_KEY).get()?.let {
-            syncManager.onSyncResponseMessage(msg, it)
-        }
+        chainSynchronizer.latestGenesisBlockResponse(msg.genesisBlockMessage)
+        log.debug("SyncResponseMessage FROM ${ctx.channel().remoteAddress()}")
     }
 
 }